production-center/core-application/apis/dv360_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pandas==2.2.2
PyYAML==6.0.2
google-cloud-secret-manager==2.20.1
google-cloud-bigquery==3.25.0
google-auth==2.34.0
google-api-python-client==2.149.0
google-cloud-storage==2.18.2
